files mv, rm

+++ Sources/ipfs-api/ipfs-call-api/FileReadDownload.swift
@@ -7,8 +7,7 @@
 
 import Foundation
 
-@available(macOS 12.0, *)
-public func FilesReadDownload(filepath: String, file:URL, completion: @escaping (Error?) -> Void ) async throws {
+public func FilesReadDownload(filepath: String, file:URL, completion: @escaping (Error?) -> Void ) throws -> Progress {
     var req:Request
     
     req = LocalRequest(path: "files/read")
@@ -17,37 +16,49 @@ public func FilesReadDownload(filepath: String, file:URL, completion: @escaping 
     
     do {
         let request:URLRequest = try req.build()
-    
+        
         let task = URLSession.shared.downloadTask(with: request) {
-                (tempURL, response, error) in
-                // Early exit on error
-                guard let tempURL = tempURL else {
-                    completion(error)
-                    return
-                }
-
-                do {
-                    // Remove any existing document at file
-                    if FileManager.default.fileExists(atPath: file.path) {
-                        try FileManager.default.removeItem(at: file)
-                    }
-
-                    // Copy the tempURL to file
-                    try FileManager.default.copyItem(
-                        at: tempURL,
-                        to: file
-                    )
-
-                    completion(nil)
-                }
-
-                // Handle potential file system errors
-                catch {
-                    completion(error)
-                }
+            (tempURL, response, error) in
+            // Early exit on error
+            guard let tempURL = tempURL else {
+                completion(error)
+                return
             }
-
-            task.resume()
+            
+            guard let response = response as? HTTPURLResponse else {
+                completion(RequestError.InvalidURLResponse)
+                return
+            }
+            
+            guard response.statusCode == 200   else {
+                completion(RequestError.UnExpectedResponseStatus(response.statusCode, response.description))
+                return
+            }
+            
+            do {
+                // Remove any existing document at file
+                if FileManager.default.fileExists(atPath: file.path) {
+                    try FileManager.default.removeItem(at: file)
+                }
+                
+                // Copy the tempURL to file
+                try FileManager.default.copyItem(
+                    at: tempURL,
+                    to: file
+                )
+                
+                completion(nil)
+            }
+            
+            // Handle potential file system errors
+            catch {
+                completion(error)
+            }
+        }
+        
+        task.resume()
+        
+        return task.progress
     } catch {
         throw error
     }
